refactor(doctor): split resource diagnostics
